Add files via upload
This is a new version of the main file for us to work out of.

It has (blank) functions that represent each of the things we need to implement in order for the overall project to work. If I've missed anything important, let me know, there's a few little things I'll be adding throughout the day.