Finalish version
Lots of changes, big refactoring.

Change to measurement mode with 'm'. Click on contours with the left mouse button. Measurements (in pixels) will be automatically generated. Toggle measurement mode on and off to take new measurements. To use the test image instead of the webcam, change the use_loaded_image variable to True